+++ fcli-core/fcli-fod/src/main/java/com/fortify/cli/fod/dast_scan/helper/FoDScanConfigDastAutomatedDescriptor.java
@@ -25,7 +25,7 @@ import lombok.ToString;
 @EqualsAndHashCode(callSuper=false)
 public class FoDScanConfigDastAutomatedDescriptor extends JsonNodeHolder {
     private Integer assessmentTypeId;
-    private Integer entitlementId;
+    private Long entitlementId;
     private String entitlementFrequencyType;
     private String scanType;
 }
